database operators/postgres-operator/postgresoperator/metrics.py
import psycopg2
from prometheus_client import CollectorRegistry, Gauge, push_to_gateway
import time

def collect_metrics(cluster_name, namespace):
    # In a real implementation, this would connect to PostgreSQL
    # and collect actual metrics. For demo purposes, we simulate.
    
    # Simulated metrics - replace with actual collection
    metrics = {
        'qps': get_qps(cluster_name, namespace),
        'latency': get_avg_latency(cluster_name, namespace),
        'connections': get_active_connections(cluster_name, namespace),
        'timestamp': time.time()
    }
    
    # Push metrics to Prometheus
    push_metrics_to_prometheus(metrics, cluster_name)
    
    return metrics

def get_qps(cluster_name, namespace):
    # TODO: Implement actual QPS collection from PostgreSQL
    return 1250  # Simulated value

def get_avg_latency(cluster_name, namespace):
    # TODO: Implement actual latency collection
    return 75  # ms, simulated value

def get_active_connections(cluster_name, namespace):
    # TODO: Implement actual connection count collection
    return 45  # Simulated value

def push_metrics_to_prometheus(metrics, cluster_name):
    registry = CollectorRegistry()
    
    g = Gauge('postgres_qps', 'Queries per second', ['cluster'], registry=registry)
    g.labels(cluster=cluster_name).set(metrics['qps'])
    
    g = Gauge('postgres_latency', 'Average query latency', ['cluster'], registry=registry)
    g.labels(cluster=cluster_name).set(metrics['latency'])
    
    g = Gauge('postgres_connections', 'Active connections', ['cluster'], registry=registry)
    g.labels(cluster=cluster_name).set(metrics['connections'])
    
    push_to_gateway('prometheus-gateway:9091', job='postgres-operator', registry=registry)
