database operators/postgres-operator/requirements.txt
kopf>=1.35
kubernetes>=18.20
prometheus-client>=0.11.0
psycopg2-binary>=2.9.3
pyyaml>=5.4.1
python-dotenv>=0.19.0
